:camera_flash: Render jupyter notebook cells
Now that the conference is over, might as well show the rendered version for people to see the figures directly without running the jupyter notebooks. Will tag a release and upload the ODG/PNG/PDF version of the poster in a bit. Also updated main README.md to include more links and an nbviewer button.