:dizzy: Use fes colourmap with black and white bathymetry
Update the offshore colours of the Antarctic-wide active subglacial lake map to be black and white using the 'fes' colourmap (part of GMT 6.2.0, added in https://github.com/GenericMappingTools/gmt/pull/4838). The previous 'oleron' wasn't too bad, but it's better to have just the grounded ice parts in colour so as to place the focus on where the action is - the active subglacial lakes! Also updated subglacial lake outlines to DeepIceDrain v0.4.1 version, and used PyGMT's `inset` function wrapped in v0.3.0!